requirements.txt
# Core Dependencies
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0

# AI / NLP
openai==1.3.0

# Payment Processing
stripe==7.4.0

# Notifications
twilio==8.10.0
sendgrid==6.11.0
firebase-admin==6.3.0

# Web Scraping (Advanced)
selenium==4.15.2
playwright==1.40.0
scrapy==2.11.0

# Data Processing
pandas==2.1.3
python-dateutil==2.8.2

# Environment Variables
python-dotenv==1.0.0

# API Framework (if building API)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database ORM (optional)
sqlalchemy==2.0.23

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
phonenumbers==8.13.26
pytz==2023.3
